ocfs2_controld: Fix compilation warnings
main.c: In function ‘checkpoint_to_proto_version’:
main.c:667:3: warning: field precision specifier ‘.*’ expects argument of type ‘int’, but argument 7 has type ‘size_t’
main.c:675:3: warning: field precision specifier ‘.*’ expects argument of type ‘int’, but argument 7 has type ‘size_t’

ckpt.c: In function ‘call_ckpt_open’:
ckpt.c:167:4: warning: format ‘%llx’ expects argument of type ‘long long unsigned int’, but argument 9 has type ‘SaCkptCheckpointHandleT’
ckpt.c: In function ‘ckpt_section_store’:
ckpt.c:473:3: warning: format ‘%d’ expects argument of type ‘int’, but argument 7 has type ‘size_t’
ckpt.c: In function ‘setup_ckpt’:
ckpt.c:614:4: warning: format ‘%llx’ expects argument of type ‘long long unsigned int’, but argument 7 has type ‘SaCkptHandleT’

Acked-by: Sunil Mushran <sunil.mushran@oracle.com>

+++ ocfs2_controld/ckpt.c
@@ -167,7 +167,7 @@ static int call_ckpt_open(struct ckpt_handle *handle, int write)
 			log_debug("Opened checkpoint \"%.*s\" with handle 0x%llx",
 				  handle->ch_name.length,
 				  handle->ch_name.value,
-				  handle->ch_handle);
+				  (unsigned long long)handle->ch_handle);
 			break;
 		}
 		if ((rc != -EAGAIN) &&
@@ -470,7 +470,7 @@ int ckpt_section_store(struct ckpt_handle *handle, const char *section,
 		return -EINVAL;
 	}
 	if (data_len > CKPT_MAX_SECTION_SIZE) {
-		log_error("Error: attempt to store %d bytes in a section "
+		log_error("Error: attempt to store %lu bytes in a section "
 			  "(max is %d)",
 			  data_len, CKPT_MAX_SECTION_SIZE);
 		return -EINVAL;
@@ -612,7 +612,7 @@ int setup_ckpt(void)
 		ais_err_to_errno(error, &rc, &reason);
 		if (!rc) {
 			log_debug("Connected to CKPT service with handle 0x%llx",
-				  daemon_handle);
+				  (unsigned long long)daemon_handle);
 			break;
 		}
 		if (rc != -EAGAIN) {
